busca, pero mal, pero no tan mal

+++ src/main/java/com/thebadtouch/searchengine/repositories/DocumentRepository.java
@@ -3,5 +3,5 @@ package com.thebadtouch.searchengine.repositories;
 import com.thebadtouch.searchengine.entities.Document;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DocumentRepository extends CrudRepository<Document, Integer> {
+public interface DocumentRepository extends CrudRepository<Document, Long> {
 }
